feat: Completes the unity gain amplifier design
This completes the design of a unity gain amplifier using an operational
amplifier (op-amp). This updates the components, equipment and
circuit schematic.

Adds practical measurements and analyis of the unity gain amplifier.
Includes measurements inside the LM741 operating range and outside the
operating range to show the limitations of the op-amp in real-world
applications.

Breaking changes: None

Resolves: None