feat: Adds integral amplifier design
This commit adds a the inegral amplifier circuit to the project. This
design includes the necessary components and configurations to implement
an integral with a sawtooth waveform with a specificed frequency
and amplitude. Also included an error situation for when the input
voltage exceeds the maximum allowable range for the amplifier.

Still to do: Calculations and formulas for component values are still
missing.

Breaking changes: None

Resolves: None