feat: Adds network analysis two branches
Adds a theorietical and practical analysis of currents on a circuit
with two power supplies and a single output. It has four resistors and
the goal is to calculate the currents from both inputs and the combined
current.

Currently the code for calculating it is missing. This will be added later.

Breaking changes: None

Resolves: None